Use connections default sender instead of 'open_sender'
The open_sender function errors if the specified target does not exist.
This change moves to using the connection object default sender which is a sending link whose target address is null.
The use of this method depends on the peer supporting so-called 'anonymous relay' semantics, which our AMQP does.
The message should have the 'to' field set to the intended destination.